backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pydantic
torch
torchvision
ultralytics
opencv-python-headless
numpy
